kubeadm: Aggregate the token functionality in sane packages.  - Factor out token constants to kubeadmconstants.  - Move cmd/kubeadm/app/util/{,token/}tokens.go  - Use the token-id, token-secret, etc constants provided by the bootstrapapi package  - Move cmd/kubeadm/app/master/tokens.go to cmd/kubeadm/app/phases/token/csv.go
This refactor basically makes it possible to hook up kubeadm to the BootstrapSigner controller later on
