+++ cmd/kubeadm/app/master/BUILD
@@ -16,7 +16,6 @@ go_library(
         "manifests.go",
         "selfhosted.go",
         "templates.go",
-        "tokens.go",
     ],
     tags = ["automanaged"],
     deps = [
@@ -31,7 +30,6 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/intstr",
-        "//vendor:k8s.io/apimachinery/pkg/util/uuid",
         "//vendor:k8s.io/apimachinery/pkg/util/wait",
         "//vendor:k8s.io/client-go/kubernetes",
         "//vendor:k8s.io/client-go/pkg/api",
