+++ cmd/kubeadm/app/phases/apiconfig/clusterroles.go
@@ -24,7 +24,6 @@ import (
 	"k8s.io/client-go/pkg/api/v1"
 	rbac "k8s.io/client-go/pkg/apis/rbac/v1beta1"
 	kubeadmconstants "k8s.io/kubernetes/cmd/kubeadm/app/constants"
-	"k8s.io/kubernetes/cmd/kubeadm/app/master"
 )
 
 const (
@@ -117,7 +116,7 @@ func CreateClusterRoleBindings(clientset *clientset.Clientset) error {
 			Subjects: []rbac.Subject{
 				{
 					Kind: "Group",
-					Name: master.KubeletBootstrapGroup,
+					Name: kubeadmconstants.CSVTokenBootstrapGroup,
 				},
 			},
 		},
