+++ cmd/kubeadm/app/util/BUILD
@@ -13,20 +13,10 @@ go_library(
     srcs = [
         "error.go",
         "template.go",
-        "tokens.go",
         "version.go",
     ],
     tags = ["automanaged"],
-    deps = [
-        "//cmd/kubeadm/app/apis/kubeadm:go_default_library",
-        "//cmd/kubeadm/app/apis/kubeadm/v1alpha1:go_default_library",
-        "//cmd/kubeadm/app/preflight:go_default_library",
-        "//pkg/bootstrap/api:go_default_library",
-        "//vendor:k8s.io/apimachinery/pkg/api/errors",
-        "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
-        "//vendor:k8s.io/client-go/kubernetes",
-        "//vendor:k8s.io/client-go/pkg/api/v1",
-    ],
+    deps = ["//cmd/kubeadm/app/preflight:go_default_library"],
 )
 
 go_test(
@@ -34,15 +24,11 @@ go_test(
     srcs = [
         "error_test.go",
         "template_test.go",
-        "tokens_test.go",
         "version_test.go",
     ],
     library = ":go_default_library",
     tags = ["automanaged"],
-    deps = [
-        "//cmd/kubeadm/app/apis/kubeadm:go_default_library",
-        "//cmd/kubeadm/app/preflight:go_default_library",
-    ],
+    deps = ["//cmd/kubeadm/app/preflight:go_default_library"],
 )
 
 filegroup(
@@ -54,6 +40,9 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [":package-srcs"],
+    srcs = [
+        ":package-srcs",
+        "//cmd/kubeadm/app/util/token:all-srcs",
+    ],
     tags = ["automanaged"],
 )
