+++ cmd/kubeadm/app/util/token/tokens_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2016 The Kubernetes Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,12 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package token
 
 import (
-	"bytes"
 	"testing"
-	"time"
 
 	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
 )
@@ -151,28 +149,6 @@ func TestRandBytes(t *testing.T) {
 	}
 }
 
-func TestDiscoveryPort(t *testing.T) {
-	var tests = []struct {
-		token    *kubeadmapi.TokenDiscovery
-		expected int32
-	}{
-		{token: &kubeadmapi.TokenDiscovery{}, expected: 9898}, // should use default
-		{token: &kubeadmapi.TokenDiscovery{Addresses: []string{"foobar:1234"}}, expected: 1234},
-		{token: &kubeadmapi.TokenDiscovery{Addresses: []string{"doesnothaveport"}}, expected: 9898}, // should use default
-		{token: &kubeadmapi.TokenDiscovery{Addresses: []string{"foorbar:abcd"}}, expected: 9898},    // since abcd isn't an int, should use default
-	}
-	for _, rt := range tests {
-		actual := DiscoveryPort(rt.token)
-		if actual != rt.expected {
-			t.Errorf(
-				"failed DiscoveryPort:\n\texpected: %d\n\t  actual: %d",
-				rt.expected,
-				actual,
-			)
-		}
-	}
-}
-
 func TestBearerToken(t *testing.T) {
 	var tests = []struct {
 		token    *kubeadmapi.TokenDiscovery
@@ -188,40 +164,6 @@ func TestBearerToken(t *testing.T) {
 				rt.expected,
 				actual,
 			)
-		}
-	}
-}
-
-func TestEncodeTokenSecretData(t *testing.T) {
-	var tests = []struct {
-		token *kubeadmapi.TokenDiscovery
-		t     time.Duration
-	}{
-		{token: &kubeadmapi.TokenDiscovery{ID: "foo", Secret: "bar"}},                 // should use default
-		{token: &kubeadmapi.TokenDiscovery{ID: "foo", Secret: "bar"}, t: time.Second}, // should use default
-	}
-	for _, rt := range tests {
-		actual := encodeTokenSecretData(rt.token, rt.t)
-		if !bytes.Equal(actual["token-id"], []byte(rt.token.ID)) {
-			t.Errorf(
-				"failed EncodeTokenSecretData:\n\texpected: %s\n\t  actual: %s",
-				rt.token.ID,
-				actual["token-id"],
-			)
-		}
-		if !bytes.Equal(actual["token-secret"], []byte(rt.token.Secret)) {
-			t.Errorf(
-				"failed EncodeTokenSecretData:\n\texpected: %s\n\t  actual: %s",
-				rt.token.Secret,
-				actual["token-secret"],
-			)
-		}
-		if rt.t > 0 {
-			if actual["expiration"] == nil {
-				t.Errorf(
-					"failed EncodeTokenSecretData, duration was not added to time",
-				)
-			}
 		}
 	}
 }
